requirements.txt
asgiref==3.11.1
boto3==1.42.72
botocore==1.42.72
certifi==2026.2.25
charset-normalizer==3.4.6
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
django-rest-framework==0.1.0
django-storages==1.14.6
djangorestframework==3.17.0
djdynatable==1.0.4
djeasyview==1.0.15
docker==7.1.0
gunicorn==25.2.0
idna==3.11
jmespath==1.1.0
packaging==26.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
